chore: move convertion of Issue to OssVulnerability
