+++ src/snyk/snykOss/services/vulnerabilityCount/vulnerabilityCountProviderLS.ts
@@ -4,7 +4,7 @@ import { ILanguageClientAdapter } from '../../../common/vscode/languageClient';
 import { ITextDocumentAdapter } from '../../../common/vscode/textdocument';
 import { InlineValueText, LSPTextDocument } from '../../../common/vscode/types';
 import { IUriAdapter } from '../../../common/vscode/uri';
-import { OssFileResult, OssResultBody, OssVulnerability, convertSeverity, isResultCliError } from '../../ossResult';
+import { OssFileResult, OssResultBody, OssVulnerability, convertIssue, isResultCliError } from '../../ossResult';
 import { OssServiceLanguageServer } from '../../ossServiceLanguageServer';
 import { ImportedModule, ModuleVulnerabilityCount, SeverityCounts } from './importedModule';
 
@@ -190,29 +190,7 @@ export class ModuleVulnerabilityCountProviderLS {
             resultCache.set(issue.filePath, res);
           }
 
-          const tempVuln: OssVulnerability = {
-            id: issue.id,
-            license: issue.additionalData.license,
-            identifiers: issue.additionalData.identifiers,
-            title: issue.title,
-            description: issue.additionalData.description,
-            language: issue.additionalData.language,
-            packageManager: issue.additionalData.packageManager,
-            packageName: issue.additionalData.packageName,
-            severity: convertSeverity(issue.severity),
-            name: issue.additionalData.name,
-            version: issue.additionalData.version,
-            exploit: issue.additionalData.exploit,
-
-            CVSSv3: issue.additionalData.CVSSv3,
-            cvssScore: issue.additionalData.cvssScore,
-
-            fixedIn: issue.additionalData.fixedIn,
-            from: issue.additionalData.from,
-            upgradePath: issue.additionalData.upgradePath,
-            isPatchable: issue.additionalData.isPatchable,
-            isUpgradable: issue.additionalData.isUpgradable,
-          };
+          const tempVuln = convertIssue(issue);
           res.vulnerabilities.push(tempVuln);
         }
       }
